+++ test/3250-ui_non_multiple_option_window_open_and_close_test.dart
@@ -174,14 +174,13 @@ void main() {
         th2Controller.triggerOptionsListRedraw();
         await tester.pumpAndSettle();
 
-        /// TODO
         // check MPTileWidget background color equals
         // theme.colorScheme.tertiaryFixed, i.e., is set.
-        // final Color setExpectedColor = Theme.of(
-        //   tester.element(editWidgetFinder),
-        // ).colorScheme.tertiaryFixed;
+        final Color setExpectedColor = Theme.of(
+          tester.element(editWidgetFinder),
+        ).colorScheme.tertiaryFixed;
 
-        // expect(mpTileWidgetID.backgroundColor, setExpectedColor);
+        expect(mpTileWidgetID.backgroundColor, setExpectedColor);
 
         // Verify the MPIDOptionWidget is gone after pressing Ok
         expect(find.byType(MPIDOptionWidget), findsNothing);